requirements.txt
gurobipy
tkinter
customtkinter
os
numpy
matplotlib
networkx
PyQt5
pandas
Pillow